landValuePredictionSystem
In the modern days, Real estate businesses have been becomes one of
major profitable industry in world. The value of a real estate can be
differs from each other based on various requirements. Location of the
land and the land surrounding facilities are the major concerns in
valuing a land. The affection of those concerns towards a land can be
changed depends on the condition. When the demand for the land
increases, the value of the land will become higher and vice verse. In
these days the real estate seller or the broker will estimate the price
of the land based on their favorable. Most of the times in real estate
businesses can be unfair to the buyer. There are no any standard
procedures of pricing a real estate. Therefore suggesting a proper
mechanism of calculating the price of a real estate would be more
effective on all the involving parties of the process. Therefore the
attempt of this chapter is to identify the main concerns of valuing a
real estate and suggesting a perfect, efficient and accurate mechanism
of pricing the real estates.